+++ packages/example/src/index.ts
@@ -1,9 +1,26 @@
-import { createMonitor, createResourceListener, createErrorListener } from '@web-resource-monitor/core'
+import { createMonitor, getResourceConfigDefault, createResourceListener, createErrorListener } from 'web-resource-monitor/src/index.ts'
 
-const monitor = createMonitor()
+
+const {fileMatcherDefault, resourceTimeoutConfigDefault} = getResourceConfigDefault()
+const monitor = createMonitor({
+  resourceListenerConfig: {
+    resourceTimeoutConfig: {
+      ...resourceTimeoutConfigDefault,
+      lll: 10
+    },
+    fileMatcher: {
+      ...fileMatcherDefault,
+      lll: ['js']
+    }
+  }
+})
 monitor.start()
 
 monitor.resourceListener.on('loaded', (a, b)=>{
+  console.log(a, b)
+})
+
+monitor.resourceListener.on('loadedTimeout', (a, b)=>{
   console.log(a, b)
 })
 
@@ -11,8 +28,13 @@ monitor.errorListener.on('error', (a, b)=>{
   console.log(a, b)
 })
 
-// setTimeout(()=>{ AAA }, 3000)
+setTimeout(()=>{ AAA }, 3000)
 
+const scriptdom = document.createElement('script')
+
+scriptdom.setAttribute('src', 'http://code.jquery.com/jquery-1.11.0.min.js')
+
+document.body.appendChild(scriptdom)
 
 // const resourceListener = createResourceListener()
 // resourceListener.start()