+++ packages/core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # web-resource-monitor
 
+## 1.3.0
+
+### Minor Changes
+
+- feat: resource use getResourceConfigDefault get default
+
+### Patch Changes
+
+- Updated dependencies
+  - @web-resource-monitor/resource@1.3.0
+  - @web-resource-monitor/error@1.3.0
+
+## 1.2.2
+
+### Patch Changes
+
+- fix resource repeated capture
+- Updated dependencies
+  - @web-resource-monitor/resource@1.2.2
+  - @web-resource-monitor/error@1.2.2
+
 ## 1.2.1
 
 ### Patch Changes