+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web-resource-monitor/shared
 
+## 1.3.0
+
+### Minor Changes
+
+- feat: resource use getResourceConfigDefault get default
+
+## 1.2.2
+
+### Patch Changes
+
+- fix resource repeated capture
+
 ## 1.2.1
 
 ### Patch Changes